public/javascript/media.js: Refactored; only show current colour scheme:
This means that the PhotoSwipe gallery that opens only contains the
gallery items matching the colour scheme of the link that was clicked.
I.e. if you click one of the light scheme links, it will only show and
cycle through light scheme items, and the same for dark colour scheme
links.